Add unit tests (#5)

+++ src/jpeg-utils/jpeg-utils.c
@@ -11,7 +11,7 @@ int readImage(const char *path, ImageData *img)
 
 	FILE *fptr = NULL;
 	int width = 0, height = 0;
-	unsigned long imgSize = 0;
+	long imgSize = 0;
 	unsigned char nComponent = 3, *compImg = NULL, *decompImg = NULL;
 	tjhandle jpegDecompressor = NULL;
 
@@ -40,7 +40,7 @@ int readImage(const char *path, ImageData *img)
 	retval = tjDecompressHeader(jpegDecompressor, compImg, imgSize, &width, &height);
 	if (retval < 0)
 		goto cleanup;
-	decompImg = (unsigned char *)malloc(width * height * nComponent * sizeof(unsigned char));
+	decompImg = (unsigned char *)malloc((size_t)width * height * nComponent * sizeof(unsigned char));
 	if (!decompImg)
 		goto cleanup;
 	retval = tjDecompress(jpegDecompressor, compImg, imgSize, decompImg, width, 0, height, nComponent, TJFLAG_FASTDCT);
